Shell script to load one file at a time
Instead of running `find` through SQL, which was having permission
issues. The shell script copies the files to /tmp/ again because of
Postgres having permission issues on the external filesystem.

I'm not enough of a Postgres config admin to figure out the permission
stuff, so this works around it, and is honestly more sensible in a lot
of ways anyway than calling programs from SQL.