Return to using Makefile + proper extensions
WARNING: This kind of edit to Flyway migrations is incompatible with any
existing DB. You'll need to drop the booksearch DB and start from
scratch if the other migration history was applied.

Applying the extensions with Flyway was intended for RDS, and we've
split that into a separate repository. So for the demonstration, let's
just use extensions where we can properly install them, drop them, etc.

There are problems that need to be solved with updating the extensions
with Flyway, and it's clearest to deal with those only for the RDS case.